Fix for RemoveShortCuts. If creation did fail it does not exist hence it cannot be removed.

+++ ETWAnalyzer/Extractors/SymbolPaths.cs
@@ -60,7 +60,11 @@ namespace ETWAnalyzer.Extract
         {
             foreach(string shortcut in myShortCuts)
             {
-                Directory.Delete(shortcut);
+
+                if (Directory.Exists(shortcut))
+                {
+                    Directory.Delete(shortcut);
+                }
             }
 
             myShortCuts.Clear();
